refactor : 자산분석 컴포넌트 분리
